Extracted validity module, fixed tests
